move block operations
